extracted logic to service to avoid multiple API requests
